requirements.txt
django
black
django-ninja